requirements.txt
pydantic
pydantic[email]
python-dotenv
requests
ipykernel
beautifulsoup4
mysql-connector-python